Move jest-dev-server to global
